Move utils to base
